API

The FittingData type

FittingObjectiveFunctions.FittingData — Type

mutable struct FittingData

Data type for fitting data.

This struct is only a container to check consistency and is not performance relevant, hence the mutability.

Fields

independent: Array of data points of the independent variable.

dependent: Array of data points of the dependent variable.

errors: Array of measurement errors of the dependent variable.

distributions: Distribution for the dependent variable uncertainty. Can be a function or an array of functions (one for each data point).

Elements with the same index belong together, i.e. define a measurement:

(independent[i], dependent[i], errors[i], distributions[i])

Constructors

FittingData(X,Y)
FittingData(X,Y,ΔY;distributions = normal_distribution)

Distributions

The distributions must have the signature (y,m,Δy), where y is the dependent data point, m is the result of the model function, given the dependent data point and the parameter, and Δy is the error of the dependent data point. By default, a normal distribution is used:

(y,m,Δy) -> exp(-(y-m)^2/(2*Δy^2))/(sqrt(2*pi) * Δy)

The ModelFunctions type

FittingObjectiveFunctions.ModelFunctions — Type

mutable struct ModelFunctions

Mutable type to collect model functions (and the respective partial derivatives) to construct objective functions from.

This struct is only a container to check consistency and is not performance relevant, hence the mutability.

Fields

model: The model function. Must have the signature (x,λ), where x is an independent data point, and λ are the parameters (e.g. an array for multiple parameters).

partials: Array of partial derivative functions (one for each direction). Must have the same signature (x,λ) as the model function.

Constructor

ModelFunctions(model, partials = nothing)

Examples

ModelFunctions((x,λ)-> λ*x)	

ModelFunctions((x,λ)-> λ*x, partials = [(x,λ)-> x])	

ModelFunctions((x,λ)-> λ[1]*x+λ[2], partials = [(x,λ)-> x, (x,λ)-> 1])

FittingObjectiveFunctions.consistency_check — Function

consistency_check(fitting_data::FittingData,model::ModelFunctions)

Test fitting_data and model, e.g. after mutation.

consistency_check(fitting_data::FittingData,model::ModelFunctions,λ)

Test if all functions can be evaluated with the parameter λ. Also, test fitting_data and model, e.g. after mutation.

Least squares objective

FittingObjectiveFunctions.lsq_objective — Function

lsq_objective(data::FittingData,model::ModelFunctions)

Return the least squares objective function lsq(λ).

Analytical expression

independent data points x_i

dependent data points y_i

errors Δy_i (defaulting to 1)

model function m

\[\text{lsq}(\lambda) = \sum_{i=1}^N \frac{(y_i - m(x_i,\lambda))^2}{\Delta y_i^2}\]

FittingObjectiveFunctions.lsq_partials — Function

lsq_partials(data::FittingData,model::ModelFunctions)

Return the partial derivatives w.r.t. the parameters [∂_1 ob(λ),…,∂_n ob(λ)] of the least squares objective function ob(λ).

The partial derivatives (∂_μ m)(x,λ) of the model function must be specified in the ModelFunctions object model.

Analytical expression

independent data points: x_i

dependent data points: y_i

errors: Δy_i (defaulting to 1)

model function: m

partial derivatives of model function in λ: ∂_μ m

\[\partial_\mu \text{lsq}(\lambda) = \sum_{i=1}^N \frac{ 2 \cdot (m(x_i,\lambda) - y_i) \cdot (\partial_\mu m)((x_i,\lambda)) }{\Delta y_i^2}\]

FittingObjectiveFunctions.lsq_gradient — Function

lsq_gradient(data::FittingData,model::ModelFunctions)

Return the gradient function grad!(gradient,λ) for the least squares objective function ob(λ). The gradient function grad! mutates (for performance) and returns the gradient. The elements of gradient do not matter, but the type and length must fit.

The partial derivatives (∂_μ m)(x,λ) of the model function must be specified in the ModelFunctions object model.

Analytical expression

independent data points: x_i

dependent data points: y_i

errors: Δy_i (defaulting to 1)

model function: m

partial derivatives of model function in λ: ∂_μ m

\[\nabla \text{lsq}(\lambda) = \sum_{\mu} \sum_{i=1}^N \frac{ 2 \cdot (m(x_i,\lambda) - y_i) \cdot (\partial_\mu m)((x_i,\lambda)) }{\Delta y_i^2} \ \vec{e}_\mu\]

Posterior objective

FittingObjectiveFunctions.posterior_objective — Function

posterior_objective(data::FittingData, model::Function,distribution::Function, prior = λ-> 1)

Return the unnormalized posterior density function p(λ).

Using the default prior λ-> 1, e.g. py passing only the first two arguments, leads to the likelihood objective for a maximum likelihood fit.

Analytical expression

independent data points x_i

dependent data points y_i

errors Δy_i (defaulting to 1)

model function m

distributions: p_i

prior distribution: p_0

\[p(\lambda) = p_0(\lambda) \cdot \prod_{i=1}^N p_i(y_i,m(x_i,\lambda),\Delta y_i)\]

FittingObjectiveFunctions.log_posterior_objective — Function

log_posterior_objective(data::FittingData,model::ModelFunctions, log_prior::Function = log_uniform_prior)

Return the logarithmic posterior density function L_p(λ). The model distributions and log-prior are assumed to be already in logarithm form.

Using the default prior, e.g. py passing only the first two arguments, leads to the logarithmic likelihood objective for a maximum likelihood fit.

Analytical expression

independent data points x_i

dependent data points y_i

errors Δy_i (defaulting to 1)

model function m

logarithmic distributions: L_i

logarithmic prior distribution: L_0

\[L_p(\lambda) = L_0(\lambda) + \sum_{i=1}^N L_i(y_i,m(x_i,\lambda),\Delta y_i)\]

FittingObjectiveFunctions.log_posterior_partials — Function

log_posterior_partials(data::FittingData,model::ModelFunctions, log_distribution_derivatives, prior_partials::Union{Nothing,AbstractArray{F,N}} = nothing) where {F <: Function, N}

Return the partial derivatives w.r.t. the parameters [∂_1 L_p(λ),…,∂_n L_p(λ)] of the log-posterior function L_p(λ).

The partial derivatives (∂_μ m)(x,λ) of the model function must be specified in the ModelFunctions object model.

The log_distribution_derivatives can either be a function (same derivative for all i), or an array of functions (∂_m L_i)(y,m,Δy)

The prior_partials can either be a nothing (for the log-likelihood), or an array of functions (∂_μ L_0)(λ).

Analytical expression

independent data points x_i

dependent data points y_i

errors Δy_i (defaulting to 1)

model function m

logarithmic distributions: L_i

logarithmic prior distribution: L_0

partial derivatives of model function in λ: ∂_μ m

partial derivatives of the log-distributions in m: ∂_m L_i

partial derivatives of the log-prior in λ: ∂_μ L_0

\[\partial_\mu L_p(\lambda) = \partial_\mu L_0(\lambda) + \sum_{i=1}^N (\partial_m L_i)(y_i, m(x_i,\lambda), \Delta y_i)\cdot (\partial_\mu m)(x_i,\lambda)\]

FittingObjectiveFunctions.log_posterior_gradient — Function

log_posterior_gradient(data::FittingData,model::ModelFunctions, log_distribution_derivatives, prior_partials::Union{Nothing,AbstractArray{F,N}} = nothing) where {F <: Function, N}

Return the gradient function grad!(gradient,λ) for the log-posterior function L_p(λ). The gradient function grad! mutates (for performance) and returns the gradient. The elements of gradient do not matter, but the type and length must fit.

The partial derivatives (∂_μ m)(x,λ) of the model function must be specified in the ModelFunctions object model.

The log_distribution_derivatives can either be a function (same derivative for all i), or an array of functions (∂_m L_i)(y,m,Δy)

The prior_gradient can either be a nothing (for the log-likelihood), or a gradient mutating function of functions ∇L_0!(gradient,λ).

Analytical expression

independent data points x_i

dependent data points y_i

errors Δy_i (defaulting to 1)

model function m

logarithmic distributions: L_i

logarithmic prior distribution: L_0

partial derivatives of model function in λ: ∂_μ m

partial derivatives of the log-distributions in m: ∂_m L_i

resulting/mutated log-prior gradient in λ: ∇ L_0

\[\nabla L_p(\lambda) = \nabla L_0(\lambda) + \sum_{\mu} \sum_{i=1}^N (\partial_m L_i)(y_i, m(x_i,\lambda), \Delta y_i)\cdot (\partial_\mu m)(x_i,\lambda) \ \vec{e}_\mu\]